Implement const expression scale and other required tasks
Summary:
- split the const expression tests up (T6723)
- add const expression scale class, and tests (T6717)
- add scale op to builder class (T6713)
- change name of const expression cast class to match other class names
- add cast (T -> T) case to const cast class (T6728)
- change "scale" from int to float in ScaleOp class
- change opmanager logic to better handle opset 0 (experimental) operators
- correctly set the type of the output on const expr cast (T6730)
- format all

Test Plan: all tests pass (OS/X). 3 new tests added

Reviewers: davidn, evan, george, simonl

Reviewed By: davidn

Differential Revision: https://phabricator.sourcevertex.net/D9085

+++ willow/src/op/scale.cpp
@@ -56,7 +56,7 @@ static OpCreator<ScaleOp> scaleOpCreator(
     [](const OperatorIdentifier &_opid,
        const Op::Settings &settings,
        const Attributes &attr) -> std::unique_ptr<Op> {
-      int64_t scale = attr.getAttribute<Attributes::Int>("scale", 1);
+      float scale = attr.getAttribute<Attributes::Float>("scale", 1.0f);
 
       return std::unique_ptr<Op>(new ScaleOp(_opid, scale, settings));
     },
